==> notebooks/nb2_data_cleaning.ipynb <==
# Notebook 2: Data Cleaning
---

import datetime as dt
import numpy as np
import pandas as pd
import warnings
warnings.filterwarnings('ignore')

## Weather Data
---

weather = pd.read_csv('../raw_data/New York City,USA 2018-06-01 to 2023-05-31.csv')

weather.info()

weather.head()

# Review summary statistics
weather.describe()

# Drop extra variables
weather.drop(['name', 'feelslike', 'dew', 'precipprob', 'preciptype', 'snow', 'snowdepth', 'windgust', 'winddir', 
              'sealevelpressure', 'cloudcover', 'solarradiation', 'solarenergy', 'uvindex', 'severerisk', 
              'conditions', 'icon', 'stations'], axis=1, inplace=True)

weather.head()

weather.isna().sum()

weather.info()

# Convert data types
weather['datetime'] = pd.to_datetime(weather['datetime'])

# Set the datetime column as the index
weather.set_index('datetime', inplace=True)

weather.isna().sum()

# Save the clean weather data
weather.to_parquet('../clean_data/clean_weather.parquet')

## Holiday Data
---

from azureml.opendatasets import PublicHolidays

# Load the holiday data
start_date = dt.datetime(2013, 6, 1)
end_date = dt.datetime.today()
hol = PublicHolidays(start_date=start_date, end_date=end_date)
holiday = hol.to_pandas_dataframe()

# Review the data structure
holiday.info()

# Review the DataFrame
holiday.head()

# Remove observations not from the US
holiday = holiday[holiday['countryRegionCode'] == 'US']

# Reset the index
holiday.reset_index(drop=True, inplace=True)

# Drop the country related columns and redundant columns
holiday.drop(['countryOrRegion', 'countryRegionCode', 'normalizeHolidayName'], axis=1, inplace=True)

# Review the data structure
holiday.info()

# Review the DataFrame
holiday.head()

# Find na values
holiday[holiday['isPaidTimeOff'].isna()]

>Martin Luther King Jr. Day became a federally paid holiday in 1983
>
>Source:
>
>https://constitutioncenter.org/blog/how-martin-luther-king-jr-s-birthday-became-a-holiday-3#:~:text=King's%20birthday%20was%20finally%20approved,state%20government%20holiday%20by%202000.

# Impute True for na values
holiday['isPaidTimeOff'][holiday['holidayName'] == 'Martin Luther King Jr. Day'] = True

# Ensure no na values exist
holiday.isna().sum()

# Convert the date column to datetime
holiday['date'] = pd.to_datetime(holiday['date'])

# Create an isHoliday column
holiday['isHoliday'] = 1

# Convert isPaidTimeOff and isHoliday to uint8 type
for col in ['isPaidTimeOff', 'isHoliday']:
    holiday[col] = holiday[col].astype('uint8')

holiday.set_index('date', inplace=True)

# Resample the data to daily frequency without forward filling
holiday = holiday.resample('D').apply(lambda x: x if x.values != np.nan else np.nan)

# Fill missing holidayName values
holiday['holidayName'].fillna('Non-Holiday', inplace=True)

# Fill remaining missing values with zero
holiday.fillna(0, inplace=True)

# Resample by hour
holiday = holiday.resample('H').ffill()

holiday = holiday.loc[dt.datetime(2018, 6, 1): dt.datetime(2023, 5, 31, 23)]

# Save the clean data to CSV
holiday.to_parquet('../clean_data/holidays.parquet')

## Citi Bike Trip Data
---

df = pd.read_parquet('../clean_data/clean_tripdata.parquet')

# Review the data structure
df.info()

# Review the DataFrame
df.head()

# Get usertype dummy variables
df[['nyc_trips_casual', 'nyc_trips_member']] = pd.get_dummies(df['usertype'])

# Drop the user column
df.drop('usertype', axis=1, inplace=True)

# Create borough start and end count columns
for col in ['start', 'end']:
    df = pd.concat([df, pd.get_dummies(df[f'{col}_borough']).add_suffix(f"_{col}_all")], axis=1)

df.head()

# Drop the start and end borough categorical columns
df.drop(['start_borough', 'end_borough'], axis=1, inplace=True)

# Make column names lowercase
df.columns = [col.lower() for col in df.columns]

# Add a trip count column
df['nyc_trips_all'] = 1

# Break down trip counts by borough and rider type
for borough in ['brooklyn', 'manhattan', 'queens']:
    for rider in ['casual', 'member']:
        for n in ['start', 'end']:
            df[f"{borough}_{n}_{rider}"] = df[f"nyc_trips_{rider}"] * df[f"{borough}_{n}_all"]

df.head()

df.info()

### Resample by Hour
---

# Resample the data by hour
df_hourly = df.resample('H').sum()

df_hourly.info()

df_hourly.isna().sum()

df_hourly.describe().astype(int)

df_hourly.to_parquet('../clean_data/hourly_tripdata.parquet')

## NYC Bike Counts
---

df = pd.read_csv('../raw_data/Bicycle_Counts.csv')

df.head()

# Drop unnecessary variables
df.drop(['countid', 'id', 'status'], axis=1, inplace=True)

# Aggregate the counts from multiple locations in NYC
df_agg = df.groupby('date').sum()

# Convert the index to datetime
df_agg.index = pd.to_datetime(df_agg.index)

# Sort the data by date
df_agg.sort_index(inplace=True)

# Keep only relevant dates
df_agg = df_agg.loc[dt.datetime(2018, 6, 1): dt.datetime(2023, 5, 31, 23, 45)]

# Resample the data by hour
df = df_agg.resample('H').sum()

# Rename the counts column to be more descriptive
df.columns = ['nyc_bike_counts']

df.head()

df.info()

df.to_parquet('../clean_data/nyc_bike_counts.parquet')
