==> notebooks/nb3_data_aggregation.ipynb <==
# Notebook 3: Data Aggregation
---

import datetime as dt
import numpy as np
import pandas as pd
from scipy.stats import zscore
import warnings
warnings.filterwarnings('ignore')

## Load the Citi Bike, Weather, Holiday, and NYC Bike Counts Data
---

# Load the data
tripdata = pd.read_parquet('../clean_data/hourly_tripdata.parquet')
weather = pd.read_parquet('../clean_data/clean_weather.parquet')
holiday = pd.read_parquet('../clean_data/holidays.parquet')
bike_counts = pd.read_parquet('../clean_data/nyc_bike_counts.parquet')

# Rename all indexes to 'date'
for df in [tripdata, weather, holiday, bike_counts]:
    df.index.name = 'date'

## Merge the Citi Bike, Weather, Holiday, and NYC Bike Counts Data
---

# Merge the DataFrames
df = pd.merge(tripdata, weather, on='date').merge(holiday, on='date').merge(bike_counts, on='date')

# Check for missing values
df.isna().sum()

df.info()

df1 = df.drop(['holidayName', 'isPaidTimeOff', 'isHoliday'], axis=1)

# Check for outliers
zscore(df1)[(zscore(df1) < -3) | (zscore(df1) > 3)].count()

def winsorize_feature(feature, lower_percentile, upper_percentile):
    """
    Apply winsorization to a feature array.

    Args:
        feature (numpy.ndarray): The array containing the feature values.
        lower_percentile (float): The lower percentile value for winsorization.
        upper_percentile (float): The upper percentile value for winsorization.

    Returns:
        numpy.ndarray: The winsorized feature array with replaced outliers.
    """
    # Calculate the winsorizing values
    lower_value = np.percentile(feature, lower_percentile)
    upper_value = np.percentile(feature, upper_percentile)

    # Replace outliers with winsorizing values
    winsorized_feature = np.where(feature < lower_value, lower_value, feature)
    winsorized_feature = np.where(winsorized_feature > upper_value, upper_value, winsorized_feature)

    return winsorized_feature

df1.describe()

# Winsorize outliers
for col in df1.columns:
    df1[col] = winsorize_feature(df1[col], 5, 95)

df1.describe()

df[df1.columns] = df1

# Save the merged data
df.to_parquet('../clean_data/compiled_data.parquet')
