==> notebooks/nb4_feature_engineering.ipynb <==
# Notebook 4: Feature Engineering
---

import numpy as np
import pandas as pd

import warnings
warnings.filterwarnings('ignore')

# Load the data
df = pd.read_parquet('../clean_data/compiled_data.parquet')

df.info()

# Extract the hour, weekday, month, and year from the start datetime
df['hour'] = (df.index.hour)
df['weekday'] = df.index.weekday
df['month'] = df.index.month
df['year'] = (df.index.year)

# Extract the season from the month
seasons = {1: 'winter', 2: 'winter', 3: 'spring', 4: 'spring', 5: 'spring', 6: 'summer', 
           7: 'summer', 8: 'summer', 9: 'autumn', 10: 'autumn', 11: 'autumn', 12: 'winter'}
df['season'] = df['month'].map(seasons)

df['weekend'] = [0 if day < 5 else 1 for day in df.index.weekday]

>NYC weekday rush hour is from 7:30 AM to 9:30 AM and 5 PM to 7 PM
>
>Citation:
>
>Fodor's Travel. (n.d.). New York City: Tips and Etiquette. The New York Times. Retrieved June 28, 2023, from https://archive.nytimes.com/www.nytimes.com/fodors/top/features/travel/destinations/unitedstates/newyork/newyorkcity/fdrs_tips_111_2.html

def categorize_hour(day_type, hour):
    """
    Categorize an hour as rush hour or not based on the day type and hour value.

    Args:
        day_type (int): Indicator for the day type (1 for rush hour, 0 for non-rush hour).
        hour (int): The hour value (24-hour format) to be categorized.

    Returns:
        str: A string indicating whether the hour is categorized as rush hour or not.
    """
    if day_type == 1:
        return 'not rush hour'
    elif hour >= 7 and hour <= 10:
        return 'rush hour'
    elif hour >= 17 and hour <=19:
        return 'rush hour'
    else:
        return 'not rush hour'

df['hour_type'] = [categorize_hour(row['weekend'], row['hour']) for _, row in df.iterrows()]

df['bike_counts_log'] = np.log10(df['nyc_bike_counts'])

df.info()

source: https://www.investopedia.com/historical-timeline-of-covid-19-in-new-york-city-5071986

def categorize_date(date):
    if date < pd.to_datetime('2020-03-07'):
        return 'pre-pandemic'
    elif date < pd.to_datetime('2020-06-08'):
        return 'lockdown'
    elif date < pd.to_datetime('2021-04-06'):
        return 'reopening'
    elif date < pd.to_datetime('2022-01-01'):
        return 'post-vaccine'
    else:
        return 'post-pandemic'

df['pandemic_period'] = pd.Series(df.index).apply(categorize_date).values

# Review the DataFrame
df.head()

# Save the new DataFrame
df.to_parquet('../clean_data/engineered_data.parquet')
