==> notebooks/nb5_data_transformation.ipynb <==
# Notebook 5: Data Transformation
---

import numpy as np
import pandas as pd
import datetime as dt

import warnings
warnings.filterwarnings('ignore')

df = pd.read_parquet('../clean_data/engineered_data.parquet')

df.head()

# Drop the holiday name column
df = df[['nyc_trips_all', 'temp', 'humidity', 'precip', 'windspeed', 'visibility', 'isPaidTimeOff', 'isHoliday', 
         'nyc_bike_counts', 'bike_counts_log', 'hour', 'weekday', 'month', 'year', 'season', 
         'weekend', 'hour_type', 'pandemic_period']]

## Perform Label Encoding for ordinal variables
---

# Label encode the pandemic_period column
pp = {'pre-pandemic': 0, 'lockdown': 1, 'reopening': 2, 'post-vaccine': 3, 'post-pandemic': 4}
df['pandemic_period'] = df['pandemic_period'].map(pp)

# Label encode the season column
seasons = {'spring': 0, 'summer': 1, 'autumn': 2, 'winter': 3,}
df['season'] = df['season'].map(seasons)

## Perform Cyclical Encoding for time-based variables
---

df['hour_sin'] = np.sin(2 * np.pi * df['hour'] / 24)
df['hour_cos'] = np.cos(2 * np.pi * df['hour'] / 24)

df['season_sin'] = np.sin(2 * np.pi * df['season'] / 4)
df['season_cos'] = np.cos(2 * np.pi * df['season'] / 4)

## Perform One-Hot Encoding for nominal variables
---

# One-hot encode the hour type column
df['rush_hour'] = pd.get_dummies(df['hour_type'])['rush hour']

# Drop the hour type column
df.drop('hour_type', axis=1, inplace=True)

# Combine the isPaidTimeOff and isHoliday columns
df['is_holiday'] = (df['isPaidTimeOff'] + df['isHoliday']).replace(2, 1)
df.drop(['isPaidTimeOff', 'isHoliday'], axis=1, inplace=True)

df.info()

df.to_parquet('../clean_data/transformed_data.parquet')
